+++ src/main/java/step1/calculator/DivideCalculator.java
@@ -1,6 +1,6 @@
-package step1.operator;
+package step1.calculator;
 
-public class DivideOperator implements Operator {
+public class DivideCalculator implements Calculator {
     @Override
     public int calculate(int firstOperand, int secondOperand) {
         return firstOperand / secondOperand;
